+++ mola_kernel/src/interfaces/FrontEndBase.cpp
@@ -1,8 +1,15 @@
-/* -------------------------------------------------------------------------
- *   A Modular Optimization framework for Localization and mApping  (MOLA)
- * Copyright (C) 2018-2025 Jose Luis Blanco, University of Almeria
- * See LICENSE for license information.
- * ------------------------------------------------------------------------- */
+/*               _
+ _ __ ___   ___ | | __ _
+| '_ ` _ \ / _ \| |/ _` | Modular Optimization framework for
+| | | | | | (_) | | (_| | Localization and mApping (MOLA)
+|_| |_| |_|\___/|_|\__,_| https://github.com/MOLAorg/mola
+
+ Copyright (C) 2018-2025 Jose Luis Blanco, University of Almeria,
+                         and individual contributors.
+ SPDX-License-Identifier: GPL-3.0
+ See LICENSE for full license information.
+*/
+
 /**
  * @file   FrontEndBase.cpp
  * @brief  Virtual interface for SLAM front-ends
@@ -72,39 +79,6 @@ void FrontEndBase::initialize(const Yaml& cfg)
         "No 'raw_data_source' entry found in the YAML definition for a "
         "FrontEndBase: YAML contents:\n"
         << cfg << "\n");
-  }
-
-  // Optional: attach to the SLAM backend:
-  auto fnd_bckends = ExecutableBase::findService<BackEndBase>();
-  if (fnd_bckends.empty())
-  {
-    MRPT_LOG_DEBUG("No SLAM back-end found in the system.");
-  }
-  else
-  {
-    if (fnd_bckends.size() > 1)
-    {
-      MRPT_LOG_WARN(
-          "More than one SLAM back-end found in the system! "
-          "Attaching to "
-          "first one.");
-    }
-    slam_backend_ = std::dynamic_pointer_cast<BackEndBase>(fnd_bckends[0]);
-    ASSERT_(slam_backend_);
-    MRPT_LOG_INFO_FMT(
-        "Attached to SLAM backend module `%s`", slam_backend_->getModuleInstanceName().c_str());
-  }
-
-  // Optional: attach to the world model.
-  {
-    auto wms = findService<WorldModel>();
-    if (wms.size() == 1)
-    {
-      worldmodel_ = std::dynamic_pointer_cast<WorldModel>(wms[0]);
-      ASSERT_(worldmodel_);
-      MRPT_LOG_INFO_FMT(
-          "Attached to WorldModel module `%s`", worldmodel_->getModuleInstanceName().c_str());
-    }
   }
 
   // Optional: attach to the visualizer: