+++ mola_input_rawlog/CHANGELOG.rst
@@ -2,6 +2,15 @@
 Changelog for package mola_input_rawlog
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.0 (2025-10-13)
+------------------
+* Modernize copyright notice
+* Remove old code that was needed to support very old MRPT versions
+* Contributors: Jose Luis Blanco-Claraco
+
+1.9.1 (2025-07-07)
+------------------
+
 1.9.0 (2025-06-06)
 ------------------